Final working state of app, ready to submit

+++ app/src/main/java/com/example/umd_gluten_free/data/Meal.kt
@@ -1,11 +1,7 @@
 package com.example.umd_gluten_free.data
 
-import android.media.Rating
-import com.google.firebase.firestore.ServerTimestamp
-import java.util.*
-
 data class Meal(
     var location: String ?= null,
     var name: String ?= null,
-    var rating: Float
+    var rating: Long
     )
